chore(hybridcloud) Move api_gateway and normalize casing
* Move the api_gateway module under sentry.hybridcloud module
* Rename api_gateway to apigateway.
* Use hybridcloud for the test module name as we'll be converging on
  that incrementally.
