+++ .github/CODEOWNERS
@@ -142,8 +142,8 @@ yarn.lock                                                @getsentry/owners-js-de
 
 ## Hybrid Cloud
 /src/sentry/silo/                                        @getsentry/hybrid-cloud
+/src/sentry/hybridcloud                                  @getsentry/hybrid-cloud
 /src/sentry/services/hybrid_cloud/                       @getsentry/hybrid-cloud
-/src/sentry/api_gateway/                                 @getsentry/hybrid-cloud
 /src/sentry/middleware/api_gateway.py                    @getsentry/hybrid-cloud
 /src/sentry/middleware/customer_domain.py                @getsentry/hybrid-cloud
 /src/sentry/middleware/subdomain.py                      @getsentry/hybrid-cloud
