--- src/sentry/api_gateway/__init__.py
@@ -1,3 +1,0 @@
-from .api_gateway import proxy_request_if_needed
-
-__all__ = ("proxy_request_if_needed",)
