+++ src/sentry/hybridcloud/apigateway/__init__.py
@@ -1,0 +1,3 @@
+from .apigateway import proxy_request_if_needed
+
+__all__ = ("proxy_request_if_needed",)
