+++ src/sentry/middleware/api_gateway.py
@@ -5,7 +5,7 @@ from typing import Any, Callable
 from django.http.response import HttpResponseBase
 from rest_framework.request import Request
 
-from sentry.api_gateway import proxy_request_if_needed
+from sentry.hybridcloud.apigateway import proxy_request_if_needed
 
 
 class ApiGatewayMiddleware:
